extracted admin modal to component
