+++ src/components/AdminModal/AdminModal.module.scss
@@ -1,0 +1,45 @@
+.background {
+  position: absolute;
+  top: 0;
+  left: 0;
+  height: 100%;
+  width: 100%;
+  background-color: rgba(black, 0.3);
+  z-index: 10;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  .modal {
+    background-color: white;
+    border-radius: 8px;
+    overflow: hidden;
+
+    .header {
+      background-color: $grey-dark;
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      box-sizing: border-box;
+      padding: 8px 16px;
+
+      .closeButton {
+        background-color: rgba(black, 0);
+        padding: 0;
+        border: none;
+        cursor: pointer;
+        border-radius: 4px;
+        transition: 100ms ease;
+
+        &:hover {
+          background-color: rgba(white, 0.1);
+        }
+      }
+
+      * {
+        color: white;
+        font-weight: normal;
+      }
+    }
+  }
+}
