+++ src/components/AdminModal/index.ts
@@ -1,0 +1,2 @@
+import AdminModal from "./AdminModal";
+export default AdminModal;
