+++ src/components/AdminTable/FilterSortModal/FilterSortModal.module.scss
@@ -1,80 +1,34 @@
-.background {
-  position: absolute;
-  top: 0;
-  left: 0;
-  height: 100%;
-  width: 100%;
-  background-color: rgba(black, 0.3);
-  z-index: 10;
+.content {
   display: flex;
-  align-items: center;
-  justify-content: center;
+  flex-direction: column;
+  box-sizing: border-box;
+  padding: 16px;
+  gap: 8px;
 
-  .modal {
-    background-color: white;
+  .filteringList {
+    border: 1px solid rgba($grey-dark, 0.1);
     border-radius: 8px;
-    overflow: hidden;
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+    align-items: center;
+    justify-content: center;
+    width: 400px;
+    box-sizing: border-box;
+    padding: 24px;
 
-    .header {
-      background-color: $grey-dark;
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      box-sizing: border-box;
-      padding: 8px 16px;
-
-      .closeButton {
-        background-color: rgba(black, 0);
-        padding: 0;
-        border: none;
-        cursor: pointer;
-        border-radius: 4px;
-        transition: 100ms ease;
-
-        &:hover {
-          background-color: rgba(white, 0.1);
-        }
-      }
-
-      * {
-        color: white;
-        font-weight: normal;
-      }
+    p {
+      color: rgba($grey-dark, 0.5);
+      user-select: none;
     }
+  }
 
-    .content {
-      display: flex;
-      flex-direction: column;
-      box-sizing: border-box;
-      padding: 16px;
-      gap: 8px;
+  .buttons {
+    display: flex;
+    gap: 8px;
 
-      .filteringList {
-        border: 1px solid rgba($grey-dark, 0.1);
-        border-radius: 8px;
-        display: flex;
-        flex-direction: column;
-        gap: 16px;
-        align-items: center;
-        justify-content: center;
-        width: 400px;
-        box-sizing: border-box;
-        padding: 24px;
-
-        p {
-          color: rgba($grey-dark, 0.5);
-          user-select: none;
-        }
-      }
-
-      .buttons {
-        display: flex;
-        gap: 8px;
-
-        button {
-          width: 100%;
-        }
-      }
+    button {
+      width: 100%;
     }
   }
 }
